+++ Dtos/PublishConcilliationDTO.cs
@@ -2,9 +2,9 @@ using System.ComponentModel.DataAnnotations;
 
 namespace PixHub.Dtos;
 
-public class PublishConcilliationDTO(int paymentProviderId, ConcilliationDTO dto)
+public class PublishConcilliationDTO(long paymentProviderId, ConcilliationDTO dto)
 {
-  public int PaymentProviderId { get; } = paymentProviderId;
+  public long PaymentProviderId { get; } = paymentProviderId;
   public DateTime Date { get; } = dto.Date;
   public string File { get; } = dto.File;
   public string Postback { get; } = dto.Postback;
