+++ Models/PixKey.cs
@@ -4,11 +4,11 @@ using System.Text.Json.Serialization;
 
 namespace PixHub.Models;
 
-public class PixKey(string value, string type, int paymentProviderAccountId)
+public class PixKey(string value, string type, long paymentProviderAccountId)
 {
   [Key]
   [DatabaseGenerated(DatabaseGeneratedOption.Identity)]
-  public int Id { get; set; }
+  public long Id { get; set; }
   public string Value { get; set; } = value;
   public string Type { get; set; } = type;
 
@@ -16,7 +16,7 @@ public class PixKey(string value, string type, int paymentProviderAccountId)
 
   [JsonIgnore]
   public PaymentProviderAccount? PaymentProviderAccount { get; set; }
-  public int PaymentProviderAccountId { get; set; } = paymentProviderAccountId;
+  public long PaymentProviderAccountId { get; set; } = paymentProviderAccountId;
 
   public DateTime CreatedAt { get; set; } = DateTime.UtcNow;
   public DateTime UpdatedAt { get; set; } = DateTime.UtcNow;
